end to end testing for notes frontend polishment
