+++ notes-frontend/playwrightTesting/playwright.config.js
